+++ CHANGELOG.rst
@@ -30,3 +30,10 @@ Changelog
 
 * Improve repr of ImportGraph.
 * Fix bug with find_shortest_path using upstream/downstream the wrong way around.
+
+latest
+------
+* Rename get_shortest_path to get_shortest_chain.
+* Rename path_exists to chain_exists.
+* Rename and reorder the kwargs for get_shortest_chain and chain_exists.
+* Raise ValueError if modules with shared descendants are passed to chain_exists if as_packages=True.